apply lint to at-date-console
